pyproject.toml
[project]
name = "saos-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "flask-oidc>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
]

[project.scripts]
saos-project = "saos_project:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
